refactor(api): separate router from handler functions
