+++ pkg/api/release.go
@@ -5,57 +5,10 @@ import (
 	"carbide-api/pkg/objects"
 	"database/sql"
 	"encoding/json"
-	"fmt"
 	"net/http"
 
 	log "github.com/sirupsen/logrus"
 )
-
-func serveRelease(w http.ResponseWriter, r *http.Request, db *sql.DB, product_name string) {
-	var release_name string
-	release_name, r.URL.Path = ShiftPath(r.URL.Path)
-	if r.URL.Path != "/" {
-		var head string
-		head, r.URL.Path = ShiftPath(r.URL.Path)
-		switch head {
-		default:
-			http.Error(w, "Not Found", http.StatusNotFound)
-		}
-		return
-	}
-	if release_name == "" {
-		switch r.Method {
-		case http.MethodGet:
-			releaseGet(w, r, db, product_name)
-			return
-		case http.MethodPost:
-			releasePost(w, r, db, product_name)
-			return
-		case http.MethodOptions:
-			return
-		default:
-			http.Error(w, fmt.Sprintf("Expected method POST or OPTIONS, got %v", r.Method), http.StatusMethodNotAllowed)
-			return
-		}
-	} else {
-		switch r.Method {
-		case http.MethodGet:
-			releaseGet1(w, r, db, product_name, release_name)
-			return
-		case http.MethodPut:
-			releasePut1(w, r, db, product_name, release_name)
-			return
-		case http.MethodDelete:
-			releaseDelete1(w, r, db, product_name, release_name)
-			return
-		case http.MethodOptions:
-			return
-		default:
-			http.Error(w, fmt.Sprintf("Expected method POST or OPTIONS, got %v", r.Method), http.StatusMethodNotAllowed)
-			return
-		}
-	}
-}
 
 // Responds with a JSON array of all releases in the database
 //
@@ -133,7 +86,7 @@ func releasePost(w http.ResponseWriter, r *http.Request, db *sql.DB, product_nam
 // Responds with the JSON representation of a release
 //
 // Success Code: 200 OK
-func releaseGet1(w http.ResponseWriter, r *http.Request, db *sql.DB, product_name string, release_name string) {
+func releaseGetByName(w http.ResponseWriter, r *http.Request, db *sql.DB, product_name string, release_name string) {
 	var retrieved_release objects.Release
 	retrieved_release.Name = &release_name
 	parent_product, err := DB.GetProduct(db, product_name)
@@ -166,7 +119,7 @@ func releaseGet1(w http.ResponseWriter, r *http.Request, db *sql.DB, product_nam
 // Accepts a JSON payload of the updated release and responds with the new JSON object after it's been successfully updated in the database
 //
 // Success Code: 200 OK
-func releasePut1(w http.ResponseWriter, r *http.Request, db *sql.DB, product_name string, release_name string) {
+func releasePutByName(w http.ResponseWriter, r *http.Request, db *sql.DB, product_name string, release_name string) {
 	var recieved_release objects.Release
 	err := json.NewDecoder(r.Body).Decode(&recieved_release)
 	if err != nil {
@@ -214,7 +167,7 @@ func releasePut1(w http.ResponseWriter, r *http.Request, db *sql.DB, product_nam
 // Deletes the release and responds with an empty payload
 //
 // Success Code: 204 No Content
-func releaseDelete1(w http.ResponseWriter, r *http.Request, db *sql.DB, product_name string, release_name string) {
+func releaseDeleteByName(w http.ResponseWriter, r *http.Request, db *sql.DB, product_name string, release_name string) {
 	var release_to_delete objects.Release
 	release_to_delete.Name = &release_name
 	parent_product, err := DB.GetProduct(db, product_name)
